Move draw shape tool pages into their own unit
